requirements.txt
streamlit>=1.30.0
folium
streamlit-folium
vertexai
google-auth>=2.22.0
google-auth-oauthlib
pandas>=1.5.0
google-cloud-aiplatform
google-cloud-bigquery>=3.0.0
matplotlib
plotly